Fix eGPU udev rules
The udev rules were too general before this (only matching PCI address),
this lead to a bug where Hyprland would try to treat the matched device
as a DRM display, which would cause it to flicker/crash.

This change makes the udev rules more specific (by including
vendor/product).

This is yet to be tested on hm90.